SaasyAI/requirements.txt
langchain==0.0.58
openai
faiss-cpu
streamlit
streamlit-chat
pandas
